Move common parts of redis connection interface to common (#6177)
