+++ source/extensions/filters/network/common/redis/client_impl.h
@@ -1,0 +1,123 @@
+#pragma once
+
+#include <chrono>
+
+#include "envoy/config/filter/network/redis_proxy/v2/redis_proxy.pb.h"
+#include "envoy/thread_local/thread_local.h"
+#include "envoy/upstream/cluster_manager.h"
+
+#include "common/buffer/buffer_impl.h"
+#include "common/common/hash.h"
+#include "common/network/filter_impl.h"
+#include "common/protobuf/utility.h"
+#include "common/upstream/load_balancer_impl.h"
+
+#include "extensions/filters/network/common/redis/client.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace Common {
+namespace Redis {
+namespace Client {
+
+// TODO(mattklein123): Circuit breaking
+// TODO(rshriram): Fault injection
+
+class ConfigImpl : public Config {
+public:
+  ConfigImpl(
+      const envoy::config::filter::network::redis_proxy::v2::RedisProxy::ConnPoolSettings& config);
+
+  bool disableOutlierEvents() const override { return false; }
+  std::chrono::milliseconds opTimeout() const override { return op_timeout_; }
+  bool enableHashtagging() const override { return enable_hashtagging_; }
+
+private:
+  const std::chrono::milliseconds op_timeout_;
+  const bool enable_hashtagging_;
+};
+
+class ClientImpl : public Client, public DecoderCallbacks, public Network::ConnectionCallbacks {
+public:
+  static ClientPtr create(Upstream::HostConstSharedPtr host, Event::Dispatcher& dispatcher,
+                          EncoderPtr&& encoder, DecoderFactory& decoder_factory,
+                          const Config& config);
+
+  ~ClientImpl();
+
+  // Client
+  void addConnectionCallbacks(Network::ConnectionCallbacks& callbacks) override {
+    connection_->addConnectionCallbacks(callbacks);
+  }
+  void close() override;
+  PoolRequest* makeRequest(const RespValue& request, PoolCallbacks& callbacks) override;
+
+private:
+  struct UpstreamReadFilter : public Network::ReadFilterBaseImpl {
+    UpstreamReadFilter(ClientImpl& parent) : parent_(parent) {}
+
+    // Network::ReadFilter
+    Network::FilterStatus onData(Buffer::Instance& data, bool) override {
+      parent_.onData(data);
+      return Network::FilterStatus::Continue;
+    }
+
+    ClientImpl& parent_;
+  };
+
+  struct PendingRequest : public PoolRequest {
+    PendingRequest(ClientImpl& parent, PoolCallbacks& callbacks);
+    ~PendingRequest();
+
+    // PoolRequest
+    void cancel() override;
+
+    ClientImpl& parent_;
+    PoolCallbacks& callbacks_;
+    bool canceled_{};
+  };
+
+  ClientImpl(Upstream::HostConstSharedPtr host, Event::Dispatcher& dispatcher, EncoderPtr&& encoder,
+             DecoderFactory& decoder_factory, const Config& config);
+  void onConnectOrOpTimeout();
+  void onData(Buffer::Instance& data);
+  void putOutlierEvent(Upstream::Outlier::Result result);
+
+  // DecoderCallbacks
+  void onRespValue(RespValuePtr&& value) override;
+
+  // Network::ConnectionCallbacks
+  void onEvent(Network::ConnectionEvent event) override;
+  void onAboveWriteBufferHighWatermark() override {}
+  void onBelowWriteBufferLowWatermark() override {}
+
+  Upstream::HostConstSharedPtr host_;
+  Network::ClientConnectionPtr connection_;
+  EncoderPtr encoder_;
+  Buffer::OwnedImpl encoder_buffer_;
+  DecoderPtr decoder_;
+  const Config& config_;
+  std::list<PendingRequest> pending_requests_;
+  Event::TimerPtr connect_or_op_timer_;
+  bool connected_{};
+};
+
+class ClientFactoryImpl : public ClientFactory {
+public:
+  // RedisProxy::ConnPool::ClientFactoryImpl
+  ClientPtr create(Upstream::HostConstSharedPtr host, Event::Dispatcher& dispatcher,
+                   const Config& config) override;
+
+  static ClientFactoryImpl instance_;
+
+private:
+  DecoderFactoryImpl decoder_factory_;
+};
+
+} // namespace Client
+} // namespace Redis
+} // namespace Common
+} // namespace NetworkFilters
+} // namespace Extensions
+} // namespace Envoy
