+++ source/extensions/health_checkers/redis/redis.h
@@ -4,6 +4,7 @@
 
 #include "common/upstream/health_checker_base_impl.h"
 
+#include "extensions/filters/network/common/redis/client_impl.h"
 #include "extensions/filters/network/redis_proxy/conn_pool_impl.h"
 
 namespace Envoy {
@@ -21,7 +22,7 @@ public:
       const envoy::config::health_checker::redis::v2::Redis& redis_config,
       Event::Dispatcher& dispatcher, Runtime::Loader& runtime, Runtime::RandomGenerator& random,
       Upstream::HealthCheckEventLoggerPtr&& event_logger,
-      Extensions::NetworkFilters::RedisProxy::ConnPool::ClientFactory& client_factory);
+      Extensions::NetworkFilters::Common::Redis::Client::ClientFactory& client_factory);
 
   static const NetworkFilters::Common::Redis::RespValue& pingHealthCheckRequest() {
     static HealthCheckRequest* request = new HealthCheckRequest();
@@ -42,8 +43,8 @@ protected:
 private:
   struct RedisActiveHealthCheckSession
       : public ActiveHealthCheckSession,
-        public Extensions::NetworkFilters::RedisProxy::ConnPool::Config,
-        public Extensions::NetworkFilters::RedisProxy::ConnPool::PoolCallbacks,
+        public Extensions::NetworkFilters::Common::Redis::Client::Config,
+        public Extensions::NetworkFilters::Common::Redis::Client::PoolCallbacks,
         public Network::ConnectionCallbacks {
     RedisActiveHealthCheckSession(RedisHealthChecker& parent, const Upstream::HostSharedPtr& host);
     ~RedisActiveHealthCheckSession();
@@ -51,7 +52,7 @@ private:
     void onInterval() override;
     void onTimeout() override;
 
-    // Extensions::NetworkFilters::RedisProxy::ConnPool::Config
+    // Extensions::NetworkFilters::Common::Redis::Client::Config
     bool disableOutlierEvents() const override { return true; }
     std::chrono::milliseconds opTimeout() const override {
       // Allow the main Health Check infra to control timeout.
@@ -59,7 +60,7 @@ private:
     }
     bool enableHashtagging() const override { return false; }
 
-    // Extensions::NetworkFilters::RedisProxy::ConnPool::PoolCallbacks
+    // Extensions::NetworkFilters::Common::Redis::Client::PoolCallbacks
     void onResponse(NetworkFilters::Common::Redis::RespValuePtr&& value) override;
     void onFailure() override;
 
@@ -69,8 +70,8 @@ private:
     void onBelowWriteBufferLowWatermark() override {}
 
     RedisHealthChecker& parent_;
-    Extensions::NetworkFilters::RedisProxy::ConnPool::ClientPtr client_;
-    Extensions::NetworkFilters::RedisProxy::ConnPool::PoolRequest* current_request_{};
+    Extensions::NetworkFilters::Common::Redis::Client::ClientPtr client_;
+    Extensions::NetworkFilters::Common::Redis::Client::PoolRequest* current_request_{};
   };
 
   enum class Type { Ping, Exists };
@@ -89,7 +90,7 @@ private:
     return std::make_unique<RedisActiveHealthCheckSession>(*this, host);
   }
 
-  Extensions::NetworkFilters::RedisProxy::ConnPool::ClientFactory& client_factory_;
+  Extensions::NetworkFilters::Common::Redis::Client::ClientFactory& client_factory_;
   Type type_;
   const std::string key_;
 };
