+++ test/extensions/filters/network/redis_proxy/mocks.cc
@@ -15,38 +15,7 @@ namespace Extensions {
 namespace NetworkFilters {
 namespace RedisProxy {
 
-MockEncoder::MockEncoder() {
-  ON_CALL(*this, encode(_, _))
-      .WillByDefault(
-          Invoke([this](const Common::Redis::RespValue& value, Buffer::Instance& out) -> void {
-            real_encoder_.encode(value, out);
-          }));
-}
-
-MockEncoder::~MockEncoder() {}
-
-MockDecoder::MockDecoder() {}
-MockDecoder::~MockDecoder() {}
-
 namespace ConnPool {
-
-MockClient::MockClient() {
-  ON_CALL(*this, addConnectionCallbacks(_))
-      .WillByDefault(Invoke([this](Network::ConnectionCallbacks& callbacks) -> void {
-        callbacks_.push_back(&callbacks);
-      }));
-  ON_CALL(*this, close()).WillByDefault(Invoke([this]() -> void {
-    raiseEvent(Network::ConnectionEvent::LocalClose);
-  }));
-}
-
-MockClient::~MockClient() {}
-
-MockPoolRequest::MockPoolRequest() {}
-MockPoolRequest::~MockPoolRequest() {}
-
-MockPoolCallbacks::MockPoolCallbacks() {}
-MockPoolCallbacks::~MockPoolCallbacks() {}
 
 MockInstance::MockInstance() {}
 MockInstance::~MockInstance() {}
